zmg_backend/requirements.txt
# Django 核心
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# 生产服务器
gunicorn>=21.2.0

# 数据库（如果以后要换成 MySQL）
# mysqlclient>=2.2.0

# 工具库
Pillow>=10.0.0
python-magic>=0.4.27
